requirements.txt
Flask==1.1.1
Flask-DotEnv==0.1.2
Flask-Script==2.0.6
python-dotenv==0.10.3
marshmallow==3.5.2
redis==4.1.4
flask_caching==1.10.1
-e .
